Split traces.rs
