+++ libdd-trace-utils/src/span/mod.rs
@@ -7,8 +7,8 @@ pub mod v04;
 pub mod v05;
 pub mod v1;
 
-mod trace;
-pub use trace::*;
+mod projection;
+pub use projection::*;
 
 use crate::msgpack_decoder::decode::buffer::read_string_ref_nomut;
 use crate::msgpack_decoder::decode::error::DecodeError;
